@@ -3,15 +3,18 @@ import userModel from "../models/userModel.js";
 
 //add items to usercart
 const addToCart = async (req,res) => {
+    console.log(req.body)
     try {
-        let userData = await userModel.findById(req.body.userId);
+        let userData = await userModel.findOne({_id:req.body.userId});
+        console.log(userData);
         let cartData = await userData.cartData;
+        //console.log(cartData)
         if(!cartData[req.body.itemId]){
             cartData[req.body.itemId] = 1;
         }else{
             cartData[req.body.itemId] += 1;
         }
-
+        //console.log("Cart Data:", userData.cartData);
         await userModel.findByIdAndUpdate(req.body.userId, {cartData});
         res.json({
             success : true,
@@ -65,5 +68,7 @@ const getCart = async (req,res) => {
         })
     }
 }
+
+
 
 export {addToCart,removeFromCart,getCart}
