@@ -1,55 +1,56 @@
 import foodModel from "../models/foodModel.js";
-import fs from 'fs';
+import fs from "fs";
 
 //add food item
-const addFood = async (req,res) => {
-    let image_filename = `${req.file.filename}`;
+const addFood = async (req, res) => {
+  let image_filename = `${req.file.filename}`;
 
-    const food = new foodModel({
-        name : req.body.name,
-        description : req.body.description,
-        price : req.body.price,
-        category : req.body.category,
-        image : image_filename
-    })
-    try{
-        await food.save();
-        res.json({success:true, message : "Food Addded"})
-    }catch(error){
-        console.log(error)
-        res.json({success : false, message : "Error"})
-    }
+  const food = new foodModel({
+    name: req.body.name,
+    description: req.body.description,
+    price: req.body.price,
+    category: req.body.category,
+    image: image_filename,
+  });
+  try {
+    await food.save();
+    res.json({ success: true, message: "Food Added" });
+} catch (error) {
+    console.log(error);
+    res.json({ success: false, message: "Error" });
 }
+};
 
-//food list
+//list all food
 const listFood = async (req,res) => {
-    try{
-        const foods = await foodModel.find({});
+    try {
+        const food = await foodModel.find({});
         res.json({
             success : true,
-            data : foods
+            data : food
         })
-    }catch(error){
-        console.log(error);
+    } catch (error) {
+        console.log(error)
         res.json({
             success : false,
             message : "Error"
         })
     }
 }
-//remove food item
-const removeFood = async (req,res) => {
-    try{
+
+//remove food
+const removeFood = async(req,res) => {
+    try {
         const food = await foodModel.findById(req.body.id);
-        fs.unlink(`uploads/${food.image}`, () => {})
-
-        await foodModel.findByIdAndDelete(req.body.id);
+        //delete the image
+        fs.unlink(`uploads/${food.image}` , () => {}) // delete from uploads folder
+        await foodModel.findByIdAndDelete(req.body.id) // delete it by mongo db
         res.json({
             success : true,
-            message : "food removed"
+            message : "Food removed"
         })
-    }catch(error){
-        console.log(error);
+    } catch (error) {
+        console.log(error)
         res.json({
             success : false,
             message : "Error"
@@ -57,4 +58,4 @@ const removeFood = async (req,res) => {
     }
 }
 
-export {addFood,listFood,removeFood}
+export { addFood ,listFood, removeFood};
