@@ -1,68 +1,151 @@
 import orderModel from "../models/orderModel.js";
 import userModel from "../models/userModel.js";
-import Stripe from "stripe";
+import Stripe from "stripe"
 
 const stripe = new Stripe(process.env.STRIPE_SECRET_KEY);
 
-//placing user order from front-end
-//logic to placeorder
+//order from front end
+
 const placeOrder = async (req, res) => {
 
-    const frontend_url = "http://localhost:5173";
+    const frontend_url = "http://localhost:5174";
 
-  try {
-    const newOrder = new orderModel({
-      userId: req.body.userId,
-      items: req.body.items,
-      amount: req.body.amount,
-      address: req.body.address,
-    });
+    try {
+        // Save the order to the database
+        const newOrder = new orderModel({
+            userId: req.body.userId,
+            items: req.body.items,
+            amount: req.body.amount, // Amount in rupees
+            address: req.body.address
+        });
+        await newOrder.save();
 
-    await newOrder.save(); //saving 
-    await userModel.findByIdAndUpdate(req.body.userId, { cartData: {} }); //cleaning uer cart data
+        // Clear the user's cart
+        await userModel.findByIdAndUpdate(req.body.userId, { cartData: {} });
 
-    //for stripe payment
-    const line_items = req.body.item.map((item) => ({
-        price_data : {
-            currency : "inr",
-            product_data : {
-                name : item.name
+        // Prepare line items for Stripe checkout
+        const line_items = req.body.items.map((item) => ({
+            price_data: {
+                currency: "inr",
+                product_data: {
+                    name: item.name
+                },
+                unit_amount: item.price * 100, // Convert item price to paise
             },
-            unit_amount : item.price * 100 * 80 // we'll get the data in $ so, we'll convert to inr
-        },
-        quantity : item.quantity
-    }))
+            quantity: item.quantity
+        }));
 
-    //delivery charges
-    line_items.push({
-        price_data : {
-            currency : "inr",
-            product_data : {
-                name : "Delivery charges"
+        // Add delivery charges
+        line_items.push({
+            price_data: {
+                currency: "inr",
+                product_data: {
+                    name: "Delivery charges"
+                },
+                unit_amount: 1000, // Assuming delivery charge is Rs. 10, converted to paise
             },
-            unit_amount : 2*100*80 // for conversion to inr
-        }
-    })
+            quantity: 1
+        });
 
-    //session : for payment
-    const session = await stripe.checkout.sessions.create({
-        line_items : line_items,
-        mode : 'payment',
-        success_url : `${frontend_url}/verify?success=true&orderId=${newOrder._id}`, //success url
-        cancel_url : `${frontend_url}/verify?success=false&orderId=${newOrder._id}`, //cancel url
-    })
+        // Create a session for payment
+        const session = await stripe.checkout.sessions.create({
+            line_items: line_items,
+            mode: 'payment',
+            success_url: `${frontend_url}/verify?success=true&orderId=${newOrder._id}`,
+            cancel_url: `${frontend_url}/verify?success=false&orderId=${newOrder._id}`
+        });
 
-    res.json({
-        success : true,
-        session_url : session.url
-    })
-  } catch (error) {
-    console.log(error);
-    res.json({
-        success : false,
-        message : "error"
-    })
-  }
+        // Respond with the session URL
+        res.json({
+            success: true,
+            session_url: session.url
+        });
+
+    } catch (error) {
+        console.error("Error in payment", error);
+        res.json({
+            success: false,
+            message: "Error occurred during the payment process"
+        });
+    }
 };
 
-export { placeOrder };
+    const verifyOrder = async (req,res) => {
+    const {orderId, success} = req.body;
+        try {
+            if(success==="true"){
+                await orderModel.findByIdAndUpdate(orderId, {payment : true});
+                res.json({
+                    success : true,
+                    message : "Paid"
+                })
+            }else{
+                await orderModel.findByIdAndDelete(orderId);
+                res.json({
+                    success : false,
+                    message : "not Paid"
+                })
+            }
+        } catch (error) {
+            console.log(error)
+            res.json({
+                success : false,
+                message : "Error"
+            })
+        }
+    }
+
+    // users order front end
+    const userOrders = async (req,res) => {
+        try {
+            const orders = await orderModel.find({userId : req.body.userId})
+            res.json({
+                success : true,
+                data : orders
+            })
+        } catch (error) {
+            console.log(error);
+            res,json({
+                success : true,
+                data : orders
+            })
+        }
+    }
+
+    //all the orders of all the users : LIsting orders for Admin panel
+    const listOrders = async (req,res) => {
+        try {
+            const orders = await orderModel.find({});
+            res.json({
+                success : true,
+                data : orders
+            })
+        } catch (error) {
+            console.log(error);
+            res.json({
+                success : false,
+                message : "Error"
+            })
+        }
+    }
+
+
+    //updating the status
+    const updateStatus = async(req,res) => {
+        try {
+            await orderModel.findByIdAndUpdate(req.body.orderId, {status : req.body.status})
+            res.json({
+                success : true,
+                message : 'Status updated'
+            })
+        } catch (error) {
+            console.log(error)
+            res.json({
+                success : false,
+                message : "Error"
+            })
+        }
+    }
+
+export { placeOrder , verifyOrder ,userOrders,listOrders,updateStatus};
+
