@@ -1,14 +1,28 @@
-import mongoose from "mongoose";
+import mongoose  from "mongoose";
 
 const foodSchema = new mongoose.Schema({
-    name : {type : String, required : true},
-    description : {type : String, required : true},
-    price : {type : Number, required: true},
-    image : {type : String , required: true},
-    category : {type : String, required : true}
-})
+    name : {
+        type : String,
+        required : true
+    },
+    description : {
+        type : String,
+        required : true
+    },
+    price : {
+        type : Number,
+        required : true
+    },
+    image : {
+        type : String,
+        required : true
+    },
+    category : {
+        type : String,
+        required : true
+    }
+});
 
+const foodModel = mongoose.model.food || mongoose.model("food", foodSchema);
 
-const foodModel =  mongoose.model.food|| mongoose.model("food", foodSchema);
-
-export default foodModel;
+export default foodModel
