@@ -1,15 +1,37 @@
-import mongoose from "mongoose";
+import mongoose from "mongoose"
 
 const orderSchema = new mongoose.Schema({
-  userId: { type: String, required: true },
-  items: { type: String, required: true },
-  amount: { type: Number, required: true },
-  address: { type: Object, required: true },
-  status: { type: String, default: "Food Processing" },
-  data: { type: Date, default: Date.now() },
-  payment: { type: Boolean, default: false },
-});
+    userId : {
+        type : String,
+        required : true,
+    },
+    items : {
+        type : Array,
+        required : true
+    },
+    amount : {
+        type : String,
+        required : true
+    },
+    address : {
+        type : Object,
+        required : true
+    },
+    status : {
+        type : String,
+        default : "Food Processing"
+    },
+    date : {
+        type : Date,
+        default : Date.now()
+    },
+    payment : {
+        type : Boolean,
+        default : false
+    },
 
-const orderModel = mongoose.model.order || mongoose.model("order", orderSchema);
+})
 
-export default orderModel;
+const orderModel = mongoose.model.order || mongoose.model("order", orderSchema)
+
+export default orderModel
