@@ -1,13 +1,25 @@
 import mongoose from "mongoose"
 
 const userSchema = new mongoose.Schema({
-    name : {type: String, required: true},
-    email : {type: String, required : true, unique : true},
-    password : {type: String, required : true},
-    cartData : {type: Object, default: {}}
-},{minimize:false});
+    name : {
+        type : String,
+        required :true
+    }, 
+    email : {
+        type : String,
+        required :true,
+        unique : true
+    },
+    password : {
+        type : String,
+        required :true
+    },
+    cartData : {
+        type : Object,
+        default : {}
+    },
+}, {minimize : false})
 
-const userModel = mongoose.model.user || mongoose.model("user",userSchema);
-
+const userModel =  mongoose.model.user || mongoose.model("user", userSchema);
 
 export default userModel;
