@@ -1,5 +1,5 @@
 import express from "express"
-import { addToCart,removeFromCart,getCart } from "../controllers/cartController.js"
+import { addToCart,removeFromCart,getCart } from "../Controllers/cartController.js"
 import authmiddleware from "../middlewares/auth.js"
 
 const cartRouter = express.Router()
