@@ -1,22 +1,21 @@
 import express from "express";
-import { addFood, listFood, removeFood } from "../controllers/foodController.js";
+import { addFood, listFood,removeFood } from "../Controllers/foodController.js";
 import multer from "multer";
 
 const foodRouter = express.Router();
 
 //img storage engine
-const storages = multer.diskStorage({
-  destination: "uploads",
-  filename: (req, file, cb) => {
+const storage = multer.diskStorage({
+    destination: "uploads",
+    filename: (req, file, cb) => {
     return cb(null, `${Date.now()}${file.originalname}`);
-  },
+    },
 });
-const upload = multer({ storage: storages });
+const upload = multer({ storage: storage });
 
 foodRouter.post("/add", upload.single("image"), addFood);
-foodRouter.get("/list", listFood);
-foodRouter.post("/remove", removeFood);
-
+foodRouter.get("/list", listFood );
+foodRouter.post("/remove", removeFood );
 
 
 export default foodRouter;
