@@ -1,10 +1,9 @@
 import express from "express"
-import {loginUser,registerUser} from "../controllers/userController.js"
+import { loginUser,registerUser } from "../Controllers/userController.js"
 
-const userRouter = express.Router()
+const userRouter = express.Router();
 
-//post 
-userRouter.post('/register',registerUser);
-userRouter.post('/login',loginUser);
+userRouter.post("/register", registerUser)
+userRouter.post("/login",loginUser)
 
 export default userRouter;
