@@ -1,84 +1,58 @@
-import React, { useContext, useState } from "react";
+import React, { useCallback, useContext, useState } from 'react'
+import {assets} from "../../assets/assets"
 import "./Navbar.css";
-import { assets } from "../../assets/assets";
-import { Link, useLocation } from "wouter";
-import { StoreContext } from "/src/context/StoreContext.jsx";
+import { Link, useNavigate } from 'react-router-dom';
+import { StoreContext } from '../context/StoreContext';
 
-const Navbar = ({ setShowLogin }) => {
-  const [menu, setMenu] = useState("Home"); // Making the Home as default
-  const { getTotalCartAmount, token, setToken } = useContext(StoreContext);
-  const [location, setLocation] = useLocation();
+const Navbar = ({setShowLogin}) => {
 
-  const logout = () => {
-    localStorage.removeItem("token");
-    setToken("");
-    setLocation("/");
-  };
+    const [menu,setMenu] = useState("home") 
+    const {getTotalCartAmount,token,setToken} = useContext(StoreContext);    
 
-  return (
-    <div className="navbar">
-      <Link to="/">
-        <img src={assets.logo} alt="" className="logo" />
-      </Link>
-      <ul className="navbar-menu">
-        <Link
-          to="/"
-          onClick={() => setMenu("Home")}
-          className={menu === "Home" ? "active" : ""}
-        >
-          Home
-        </Link>
-        <a
-          href="#expore-menu"
-          onClick={() => setMenu("Menu")}
-          className={menu === "Menu" ? "active" : ""}
-        >
-          Menu
-        </a>
-        <a
-          href="#app-download"
-          onClick={() => setMenu("Mobile-App")}
-          className={menu === "Mobile-App" ? "active" : ""}
-        >
-          Mobile-App
-        </a>
-        <a
-          href="#footer"
-          onClick={() => setMenu("Contact-us")}
-          className={menu === "Contact-us" ? "active" : ""}
-        >
-          Contact Us
-        </a>
-      </ul>
-      <div className="navbar-right">
-        <img src={assets.search_icon} alt="" />
-        <div className="navbar-search-icon">
-          <Link to="/cart">
-            <img src={assets.basket_icon} alt="" />
-          </Link>
-          <div className={getTotalCartAmount() === 0 ? "" : "dot"}></div>
+    const navigate = useNavigate();
+
+    //just reomve token from local storage
+    const logout = () => {
+        localStorage.removeItem("token");
+        setToken("");
+        navigate("/");//redirect to home page
+    }
+    return (
+    <div className='navbar'>
+        <Link to="/"><img src = {assets.logo} alt="logo" className='logo' /></Link>
+        <ul className="navbar-menu">
+            <Link to="/" onClick= {() => { setMenu("home")}}className={menu === "home"  ? "active" : "" }>Home</Link>
+            <a href='#explore-menu' onClick= {() => { setMenu("menu")}}className={menu === "menu" ? "active" : ""}>Menu</a>
+            <a href='#app-download' onClick= {() => { setMenu("mobile-app")}}className={menu === "mobile-app"  ? "active" : "" }>Mobile-app</a>
+            <a href='#footer' onClick= {() => { setMenu("contact-us")}}className={menu === "contact-us"  ? "active" : "" }>Contact us</a>
+        </ul>
+        <div className="navbar-right">
+            <img src={assets.search_icon} alt="" />
+            <div className="navbar-search-icon">
+                <Link to="/cart"> <img src={assets.basket_icon} alt="" /></Link>
+                <div className={getTotalCartAmount() === 0  ? "":"dot"}></div>
+            </div>
+                {!token ? (
+            <button onClick={() => setShowLogin(true)}>Sign In</button>
+            ) : (
+            <div className="navbar-profile">
+                <img src={assets.profile_icon} alt="" />
+                <ul className="nav-profile-dropdown">
+                <li>
+                    <img onClick= {() => navigate('/myorders')}src={assets.bag_icon} alt="" />
+                    <p>Orders</p>
+                </li>
+                <hr />
+                <li>
+                    <img onClick = {logout} src={assets.logout_icon} alt="" />
+                    <p>Logout</p>
+                </li>
+                </ul>
+            </div>
+            )}
         </div>
-        {!token ? (
-          <button onClick={() => setShowLogin(true)}>Sign In</button>
-        ) : (
-          <div className="navbar-profile">
-            <img src={assets.profile_icon} alt="" />
-            <ul className="nav-profile-dropdown">
-              <li>
-                <img src={assets.bag_icon} alt="" />
-                <p>Orders</p>
-              </li>
-              <hr />
-              <li onClick={logout}>
-                <img src={assets.logout_icon} alt="" />
-                <p>Logout</p>
-              </li>
-            </ul>
-          </div>
-        )}
-      </div>
     </div>
-  );
-};
+    ) 
+}
 
-export default Navbar;
+export default Navbar
