@@ -1,19 +1,16 @@
-import React, { useContext } from "react";
-import "./Cart.css";
-import { StoreContext } from "../../context/StoreContext";
+import React, { useContext } from 'react'
+import "./Cart.css"
+import { StoreContext } from '../../Components/context/StoreContext'
 import { useNavigate } from 'react-router-dom';
 
-
 const Cart = () => {
-  // adding the cart functionality
-  const { food_list, cartItems, addToCart, setCartItems, removeFromCart , getTotalCartAmount, url} = useContext(StoreContext);
 
-   const navigate = useNavigate();
+  const {cartItems, food_list, removeFromCart, getTotalCartAmount, url} = useContext(StoreContext);
 
-  const DeliveryCharges = 40;
+  const navigate = useNavigate();
 
   return (
-    <div className="cart">
+    <div className='cart'>
       <div className="cart-items">
         <div className="cart-items-title">
           <p>Items</p>
@@ -46,7 +43,7 @@ const Cart = () => {
             );
           }
         })}
-      </div>
+      </div>  
       <div className="cart-bottom">
         <div className="cart-total">
           <h2>Cart Total</h2>
@@ -58,12 +55,12 @@ const Cart = () => {
             <hr />
             <div className="cart-total-details">
               <p>Delivery Fees</p>
-              <p>₹ {40}</p>
+              <p>₹ {getTotalCartAmount() === 0 ? 0 : 10 }</p>
             </div>
             <hr />
             <div className="cart-total-details">
               <b>Total</b>
-              <b>₹  {getTotalCartAmount()  + 40}</b>
+              <b>₹  {getTotalCartAmount() === 0 ? 0 : getTotalCartAmount()  + 10}</b>
             </div>
           </div>
             <button onClick={() => navigate('/order')}>PROCEED TO CHECKOUT</button>
@@ -73,13 +70,13 @@ const Cart = () => {
             <p>Got a promo code? Enter it here.</p>
             <div className="cart-promocode-input">
               <input type="text " placeholder="Promo-code" />
-              <button>Submit</button>
+              <button >Submit</button>
             </div>
           </div>
         </div>
       </div>
     </div>
-  );
-};
+  )
+}
 
-export default Cart;
+export default Cart
