@@ -1,18 +1,18 @@
-import React,{useState} from 'react'
-import './Home.css'
+import React , { useState } from 'react'
+import "./Home.css"
 import Header from '../../Components/Header/Header'
 import ExploreMenu from '../../Components/ExploreMenu/ExploreMenu'
 import FoodDisplay from '../../Components/FoodDisplay/FoodDisplay'
 import AppDownload from '../../Components/AppDownload/AppDownload'
+const Home = () => {
 
-const Home = () => { 
   const [category, setCategory] = useState("All");
 
   return (
     <div>
       <Header/>
-      <ExploreMenu category = {category} setCategory ={setCategory}/>
-      <FoodDisplay category={category}/>
+      <ExploreMenu category={category} setCategory={setCategory}/>
+      <FoodDisplay category={category}  />
       <AppDownload/>
     </div>
   )
