@@ -1,7 +1,10 @@
-import React, { useContext, useState  } from "react";
-import "./Placeorder.css";
-import { StoreContext } from "../../context/StoreContext";
+import React, { useContext, useEffect, useState } from 'react'
+import "./Placeorder.css"
+import { StoreContext } from '../../Components/context/StoreContext';
 import axios from "axios";
+import {useNavigate} from "react-router-dom"
+
+
 const Placeorder = () => {
 
   const {getTotalCartAmount,token,food_list,cartItems,url} = useContext(StoreContext);
@@ -26,28 +29,37 @@ const Placeorder = () => {
 
   const placeOrder = async (event) => {
     event.preventDefault();
-    let orderItems = [];
+    const orderItems = [];
     food_list.map((item) => {
       if(cartItems[item._id] > 0){
-        let itemInfo = item;
-        itemInfo["quantity"] = cartItems[item._id];
-        orderItems.push(itemInfo);
+          let itemInfo = item;
+          itemInfo["quantity"] = cartItems[item._id];
+          orderItems.push(itemInfo)
       }
-    })
-    let orderData = {
-      address : data,
-      items : orderItems,
-      amount : getTotalCartAmount() + 40,
-    }
-    let response = await axios.post(url + "/api/order/place",orderData, { headers : {token}});
+  })
+  let orderData = {
+    address : data,
+    items : orderItems,
+    amount : getTotalCartAmount() + 10,
+  }
+  let response = await axios.post(url + "/api/order/place", orderData , {headers : {token}})
     if(response.data.success){
       const {session_url} = response.data;
-      window.location.replace(session_url);
-     }else{
-       alert("Error");
-     }
-  }
-  
+        window.location.replace(session_url)
+    }else{
+      alert("Error in payment, try again")
+    }
+  } 
+
+  const navigate = useNavigate()
+
+  useEffect(() => {
+    if(!token){
+      navigate('/cart')
+    }else if(getTotalCartAmount() === 0){
+      navigate('/cart')
+    }
+  },[token])
 
 
   return (
@@ -61,14 +73,14 @@ const Placeorder = () => {
         <input required name="email" onChange={onChangeHandler} value={data.email} type="email" placeholder="Email Address" />
         <input required name="street" onChange={onChangeHandler} value={data.street} type="text" placeholder="Street" />
         <div className="multi-fields">
-          <input required name="city" onChange={onChangeHandler} value={data.city} type="text" placeholder="City" />
+          <input required name="city"onChange={onChangeHandler} value={data.city}  type="text" placeholder="City" />
           <input required name="state" onChange={onChangeHandler} value={data.state} type="text" placeholder="State" />
         </div>
         <div className="multi-fields">
           <input required name="zipcode" onChange={onChangeHandler} value={data.zipcode} type="text" placeholder="Zip code" />
           <input required name="country" onChange={onChangeHandler} value={data.country} type="text" placeholder="Country" />
         </div>
-        <input required name="phone" onChange={onChangeHandler} value={data.phone} type="text" placeholder="Phone" />
+        <input required name="phone"onChange={onChangeHandler} value={data.phone}  type="text" placeholder="Phone" />
       </div>
       <div className="place-order-right">
       <div className="cart-total">
@@ -81,12 +93,12 @@ const Placeorder = () => {
             <hr />
             <div className="cart-total-details">
               <p>Delivery Fees</p>
-              <p>₹ {40}</p>
+              <p>₹ {getTotalCartAmount() === 0 ? 0 : 10 }</p>
             </div>
             <hr />
             <div className="cart-total-details">
               <b>Total</b>
-              <b>₹  {getTotalCartAmount()  + 40}</b>
+              <b>₹  {getTotalCartAmount() === 0 ? 0 : getTotalCartAmount()  + 10}</b>
             </div>
           </div>
             <button type="submit">PROCEED TO PAY</button>
@@ -96,4 +108,4 @@ const Placeorder = () => {
   );
 };
 
-export default Placeorder;
+export default Placeorder
